Modularizing and adding tests (#28)
<!-- Stacky Stack Info -->
**Stack:**
- yash/modularize (#28 ❌) ← (CURRENT PR)
<!-- End Stacky Stack Info -->
