+++ src/stacky/commands/__init__.py
@@ -1,0 +1,1 @@
+# Commands module - command handlers for stacky
