+++ src/stacky/git/__init__.py
@@ -1,0 +1,1 @@
+# Git module - git operations for stacky
