+++ src/stacky/pr/__init__.py
@@ -1,0 +1,1 @@
+# PR module - GitHub PR operations
