+++ src/stacky/stack/__init__.py
@@ -1,0 +1,1 @@
+# Stack module - stack data structures and operations
